move wrapper utils to own class
